// frontend/src/app/Modulos/ventas/pages/moleculas-rv/interfaces/VentasMoleculas.interface.ts
export interface VentasMoleculas {
    UDCodes_CodeID:       string;
    UDCodes_CodeDesc:     string;
    Calculated_Venta:     number;
    Calculated_Costo:     number;
    Calculated_Peso:      number;
    Calculated_PPTO:      number;
    Calculated_Ejecutado: number;
    Calculated_Pendiente: number;
    Calculated_GAP:       number;
    Calculated_MG:        number;
    Calculated_MGp:       number;
    RowIdent:             string;
}